requirements.txt
accelerate>=0.25.0
autocrop>=1.3.0
awscli>=1.31.12
boto3>=1.33.12
bitsandbytes>=0.41.0
deepspeed>=0.12.6
diffusers==0.25.0
matplotlib>=3.8.0
peft==0.7.1
pillow>=10.0.1
prodigyopt>=1.0
safetensors>=0.4.1
sagemaker>=2.199.0
tensorboard>=2.15.1
torch==2.0.1
torchvision>=0.15.2
tqdm>=4.66.1
transformers==4.36.2
wandb>=0.16.1
xformers>=0.0.19